+++ libs/dfx-helper/src/lib/directives/ping.ts
@@ -4,15 +4,15 @@ import {HttpClient} from '@angular/common/http';
 import {coerceNumberProperty, NumberInput} from '@angular/cdk/coercion';
 import {catchError, distinctUntilChanged, map, of, switchMap, timer} from 'rxjs';
 import {interceptorByPass} from '../interceptor/by-pass-interceptor.builder';
-import {AComponent} from '../components/abstract-component';
+import {ADirective} from '../components/abstract-directive';
 
 @Component({
   selector: '[hideIfPingSucceeds]',
   standalone: true,
-  template: '',
+  template: '<ng-content></ng-content>',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class DfxHideIfPingSucceeds extends AComponent {
+export class DfxHideIfPingSucceeds extends ADirective {
   @Input() url!: string;
 
   /**
@@ -35,11 +35,14 @@ export class DfxHideIfPingSucceeds extends AComponent {
   constructor(httpClient: HttpClient, changeDetectorRef: ChangeDetectorRef) {
     super();
     this.unsubscribe(
-      timer(0, this._refreshTime)
+      timer(0, this._refreshTime * 1000)
         .pipe(
-          switchMap(() => httpClient.get(this.url, this.byPassLoggingInterceptor)),
-          map(() => true),
-          catchError(() => of(false)),
+          switchMap(() =>
+            httpClient.get(this.url, this.byPassLoggingInterceptor).pipe(
+              map(() => false),
+              catchError(() => of(true))
+            )
+          ),
           distinctUntilChanged()
         )
         .subscribe((isOffline) => {
@@ -53,7 +56,7 @@ export class DfxHideIfPingSucceeds extends AComponent {
 @Component({
   selector: '[hideIfPingFails]',
   standalone: true,
-  template: '',
+  template: '<ng-content></ng-content>',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class DfxHideIfPingFails extends DfxHideIfPingSucceeds {